Add files via upload
This is a Anaconda Jupyter notebook variation.
Since openpyxl is pre-installed on Anaconda already so there is no need to install it here.
We only need to install docx package. The !{sys.executable} is to ensure that you are running the pip version associated with the current Python kernel, so that the installed packages can be used in the current notebook. 
Please note to remember to first look at your excel file to decide how to parse using panda as column names maybe different then the default entered here.